DI orderDAO

+++ prjNorthwindCustomer/Helper/CustomersHelper.cs
@@ -7,17 +7,12 @@ namespace prjNorthwindCustomer.Helper
     internal class CustomersHelper : ICustomersHelper
     {
         private readonly CustomersDAO _customersDAO;
+        private readonly OrdersDAO _ordersDAO;
 
-        public CustomersHelper(CustomersDAO customersDAO)
+        public CustomersHelper(CustomersDAO customersDAO, OrdersDAO ordersDAO)
         {
             _customersDAO = customersDAO;
-        }
-
-        private string _connectionString { get; set; }
-
-        internal CustomersHelper(string connection)
-        {
-            _connectionString = connection;
+            _ordersDAO = ordersDAO;
         }
 
         public IEnumerable<Customers> selectCustomersInfo(Customers input = null) 
@@ -40,6 +35,10 @@ namespace prjNorthwindCustomer.Helper
             {
                 throw new InvalidOperationException("查無此顧客");
             }
+            if (_ordersDAO.hasOrder(customerID))
+            {
+                throw new InvalidOperationException("該顧客有其他訂單紀錄，請先刪除訂單紀錄後才能刪除顧客。");
+            }
             _customersDAO.deleteCustomer(customerID); 
         }
 
@@ -61,8 +60,6 @@ namespace prjNorthwindCustomer.Helper
 
         public void createNewCustomer(Customers input) 
         {
-            OrdersDAO oDAO = new OrdersDAO(_connectionString);
-
             if (_customersDAO.hasCustomer(input.CustomerID))
             {
                 throw new InvalidOperationException("顧客編號與其他人重複，請重新輸入新顧客編號");
@@ -74,11 +71,7 @@ namespace prjNorthwindCustomer.Helper
             if (input.CustomerID.Length > 5)
             {
                 throw new ArgumentException("顧客編號最多只能輸入5個字元", nameof(input.CustomerID));
-            }
-            if (oDAO.hasOrder(input.CustomerID)) 
-            {
-                throw new InvalidOperationException("該顧客有其他訂單紀錄，請先刪除訂單紀錄後才能刪除顧客。");
-            }
+            }      
             _customersDAO.createNewCustomer(input);
         }
